requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
nlpaug==1.1.11

# ML utilities
scikit-learn>=1.2.0

# Progress bars
tqdm>=4.65.0

# Experiment tracking (optional)
mlflow>=2.3.0

# Development tools (optional)
pytest>=7.3.0
black>=23.3.0
